+++ shared/src/main/java/net/blay09/mods/waystones/tag/ModItemTags.java
@@ -11,8 +11,8 @@ public class ModItemTags {
     public static final TagKey<Item> RETURN_SCROLLS = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "return_scrolls"));
     public static final TagKey<Item> WARP_SCROLLS = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "warp_scrolls"));
     public static final TagKey<Item> WARP_STONES = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "warp_stones"));
-    public static final TagKey<Item> SHARDS = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "shards"));
-    public static final TagKey<Item> SHARDS_CONSUMABLE = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "shards_consumable"));
+    public static final TagKey<Item> WARP_SHARDS = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "warp_shards"));
+    public static final TagKey<Item> SINGLE_USE_WARP_SHARDS = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "single_use_warp_shards"));
     public static final TagKey<Item> WAYSTONES = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "waystones"));
     public static final TagKey<Item> SHARESTONES = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "sharestones"));
     public static final TagKey<Item> DYED_SHARESTONES = TagKey.create(Registries.ITEM, new ResourceLocation(Waystones.MOD_ID, "dyed_sharestones"));
